# docs/api/openapi.json
{"openapi":"3.0.1","info":{"title":"Connect U","description":"유레카 미니프로젝트2","version":"v1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Users API","description":"회원 관련 API"},{"name":"Posts","description":"게시글(Post) 생성, 조회, 삭제 및 이미지 업로드 API"},{"name":"Posts","description":"게시글(Post) 좋아요 추가 및 취소 API"},{"name":"Comment API","description":"댓글 / 대댓글 / 정렬 / 페이징 관련 API"},{"name":"Follow API","description":"팔로우 관련 API"},{"name":"Auth API","description":"인증 관련 API"},{"name":"Comment Like API","description":"댓글 좋아요 관련 API"},{"name":"Search API","description":"키워드 기반 사용자 / 게시글 통합 검색 API"}],"paths":{"/api/posts/{id}":{"get":{"tags":["Posts"],"summary":"게시글 단건 조회","description":"게시글 ID를 기준으로 게시글 상세 정보를 조회합니다.","operationId":"getPost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}}}},"put":{"tags":["Posts"],"operationId":"updatePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"deleteImageSeqs","in":"query","required":false,"schema":{"type":"array","items":{"type":"integer","format":"int32"}}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}}}},"delete":{"tags":["Posts"],"summary":"게시글 삭제","description":"게시글 작성자 본인만 게시글을 삭제할 수 있습니다. (Soft Delete)","operationId":"deletePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/users/images":{"post":{"tags":["Users API"],"summary":"회원 프로필 이미지(1개) 업로드","operationId":"uploadImage","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"},"usersId":{"type":"integer","format":"int32"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/posts":{"get":{"tags":["Posts"],"summary":"게시글 목록 조회 (기본)","description":"전체 게시글을 페이징하여 조회합니다. Pageable(page, size, sort)를 지원합니다.","operationId":"listPosts","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoPostResponse"}}}}}},"post":{"tags":["Posts"],"summary":"게시글 작성","description":"로그인한 사용자가 새로운 게시글을 작성합니다.","operationId":"createPost","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}}}}},"/api/posts/{id}/images":{"post":{"tags":["Posts"],"summary":"게시글 이미지 업로드","description":"게시글에 이미지를 업로드합니다.\nmultipart/form-data 형식으로 파일을 전달해야 합니다.\n","operationId":"uploadImage_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostImageDto"}}}}}}},"/api/posts/likes/{postId}":{"post":{"tags":["Posts"],"summary":"게시글 좋아요 추가","description":"로그인한 사용자가 특정 게시글에 좋아요를 추가합니다.","operationId":"createLike","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Posts"],"summary":"게시글 좋아요 취소","description":"로그인한 사용자가 특정 게시글의 좋아요를 취소합니다.","operationId":"deleteLike","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/follow/{targetUsersId}":{"post":{"tags":["Follow API"],"summary":"팔로우","operationId":"follow","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"targetUsersId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Follow API"],"summary":"언팔로우","operationId":"unfollow","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"targetUsersId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/comments":{"post":{"tags":["Comment API"],"summary":"댓글 작성","description":"부모 댓글 ID가 없으면 일반 댓글, 있으면 대댓글로 등록됩니다.","operationId":"createComment","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentCreateRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentResponseDto"}}}}}}},"/api/comments/likes/{commentId}":{"post":{"tags":["Comment Like API"],"summary":"댓글 좋아요 토글","description":"해당 댓글에 좋아요를 누르거나, 이미 눌렀다면 좋아요를 취소합니다.","operationId":"toggleLike","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentLikeResponseDto"}}}}}}},"/api/auth/logout":{"post":{"tags":["Auth API"],"summary":"로그아웃","description":"해당 유저의 세션을 만료시킵니다.","operationId":"logout","responses":{"200":{"description":"OK"}}}},"/api/auth/login":{"post":{"tags":["Auth API"],"summary":"로그인","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}}}},"/api/auth/join":{"post":{"tags":["Auth API"],"summary":"회원가입","description":"이미지는 회원가입 후 별도의 API를 통해 업로드합니다.","operationId":"join","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JoinResponseDto"}}}}}}},"/api/users":{"delete":{"tags":["Users API"],"summary":"회원 탈퇴","description":"soft delete, 해당 유저의 세션을 무효화합니다.","operationId":"quit","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["Users API"],"summary":"회원 정보 수정","description":"정보 수정 후 해당 유저의 세션을 무효화합니다. 다시 로그인해주세요.","operationId":"update","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsersRequestDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/comments/{commentId}":{"delete":{"tags":["Comment API"],"summary":"댓글 삭제","description":"댓글은 soft delete 처리되며, 작성자 본인만 삭제할 수 있습니다.","operationId":"deleteComment","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["Comment API"],"operationId":"updateComment","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"content","in":"query","required":true,"schema":{"type":"string"}},{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentResponseDto"}}}}}}},"/api/users/my-info":{"get":{"tags":["Users API"],"summary":"자기 자신의 회원 세부 정보 조회","operationId":"getMyInfo","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsersResponseDto"}}}}}}},"/api/users/id/{usersId}":{"get":{"tags":["Users API"],"summary":"식별자로 회원 세부 정보 조회","operationId":"getUsersById","parameters":[{"name":"myUsersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"usersId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsersResponseDto"}}}}}}},"/api/search/users":{"get":{"tags":["Search API"],"summary":"키워드 기반 회원 닉네임으로 검색","operationId":"searchUsersByKeyword","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoUsersListResponseDto"}}}}}}},"/api/search/post":{"get":{"tags":["Search API"],"summary":"키워드 기반 게시글 제목/본문을 검색","operationId":"searchPostByKeyword","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoPostResponse"}}}}}}},"/api/posts/likes":{"get":{"tags":["Posts"],"summary":"게시글 전체 좋아요순 조회","description":"게시글을 좋아요 개수 기준 내림차순으로 조회합니다.","operationId":"getActivePostLikeCountDesc","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoPostResponse"}}}}}}},"/api/posts/latest":{"get":{"tags":["Posts"],"summary":"게시글 전체 최신순 조회","description":"게시글을 최신순으로 조회합니다.\npage, size를 쿼리 파라미터로 전달하세요.\n반환된 content가 빈 리스트라면 마지막 페이지입니다.\n","operationId":"getActivePostsLatest","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoPostResponse"}}}}}}},"/api/posts/following-likes":{"get":{"tags":["Posts"],"summary":"팔로우한 사용자의 게시글 좋아요순 조회","description":"팔로우 대상이 작성한 게시글을 좋아요 수 기준으로 조회합니다.","operationId":"getActiveFollowingPostsLikeCountDesc","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoPostResponse"}}}}}}},"/api/posts/following-latest":{"get":{"tags":["Posts"],"summary":"팔로우한 사용자의 게시글 최신순 조회","description":"로그인 사용자가 팔로우한 대상의 게시글을 최신순으로 조회합니다.\n조회 결과가 더 이상 없으면 게시글 전체 최신순 조회 API를 사용하세요.\n","operationId":"getActiveFollowingPostsLatest","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoPostResponse"}}}}}}},"/api/follow/my-followings":{"get":{"tags":["Follow API"],"summary":"자신이 팔로우하는 유저 목록 조회","description":"페이지 번호(page), 페이지 크기(size)를 URL 쿼리 파라미터로 전달해주세요.","operationId":"getMyFollowings","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoUsersListResponseDto"}}}}}}},"/api/follow/my-followers":{"get":{"tags":["Follow API"],"summary":"자신을 팔로우하는 유저 목록 조회","description":"페이지 번호(page), 페이지 크기(size)를 URL 쿼리 파라미터로 전달해주세요.","operationId":"getMyFollowers","parameters":[{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoUsersListResponseDto"}}}}}}},"/api/follow/followings/{usersId}":{"get":{"tags":["Follow API"],"summary":"특정 유저가 팔로우하는 유저 목록 조회","description":"페이지 번호(page), 페이지 크기(size)를 URL 쿼리 파라미터로 전달해주세요.","operationId":"getUsersFollowings","parameters":[{"name":"usersId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoUsersListResponseDto"}}}}}}},"/api/follow/followers/{usersId}":{"get":{"tags":["Follow API"],"summary":"특정 유저를 팔로우하는 유저 목록 조회","description":"페이지 번호(page), 페이지 크기(size)를 URL 쿼리 파라미터로 전달해주세요.","operationId":"getUsersFollowers","parameters":[{"name":"usersId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoUsersListResponseDto"}}}}}}},"/api/comments/{postId}":{"get":{"tags":["Comment API"],"operationId":"getCommentsByPost","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"latest"}},{"name":"usersId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagingResultDtoCommentTreeResponseDto"}}}}}}},"/api/auth/csrf":{"get":{"tags":["Auth API"],"summary":"백엔드 개발 API 테스트를 위한 CSRF 토큰 발급 API입니다.","description":"Postman 테스트 시 'X-CSRF-TOKEN' 헤더에 토큰 값을 첨부해주세요.","operationId":"csrf","parameters":[{"name":"csrfToken","in":"query","required":true,"schema":{"$ref":"#/components/schemas/CsrfToken"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CsrfToken"}}}}}}},"/api/auth/check-email/{email}":{"get":{"tags":["Auth API"],"summary":"이메일 중복 확인","operationId":"checkEmail","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"CreatePostRequest":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"}}},"PostImageDto":{"type":"object","properties":{"seq":{"type":"integer","format":"int32"},"imageKey":{"type":"string"},"imageUrl":{"type":"string"}}},"PostResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"authorId":{"type":"integer","format":"int32"},"authorUsername":{"type":"string"},"title":{"type":"string"},"content":{"type":"string"},"likeCount":{"type":"integer","format":"int32"},"isLiked":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"isDeleted":{"type":"boolean"},"deletedAt":{"type":"string","format":"date-time"},"images":{"type":"array","items":{"$ref":"#/components/schemas/PostImageDto"}},"imageKeys":{"type":"array","items":{"type":"string"}}}},"CommentCreateRequestDto":{"type":"object","properties":{"postId":{"type":"integer","format":"int32"},"parentCommentId":{"type":"integer","format":"int32"},"content":{"type":"string"}}},"CommentResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"content":{"type":"string"},"likeCount":{"type":"integer","format":"int32"},"childCount":{"type":"integer","format":"int32"},"isDeleted":{"type":"boolean"},"deletedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int32"},"username":{"type":"string"},"parentCommentId":{"type":"integer","format":"int32"},"isLiked":{"type":"boolean"}}},"CommentLikeResponseDto":{"type":"object","properties":{"commentId":{"type":"integer","format":"int32"},"likeCount":{"type":"integer","format":"int32"},"isLiked":{"type":"boolean"}}},"LoginRequestDto":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"LoginResponseDto":{"type":"object","properties":{"nickname":{"type":"string"}}},"JoinRequestDto":{"required":["email","nickname","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"nickname":{"type":"string"},"description":{"type":"string"}}},"JoinResponseDto":{"type":"object","properties":{"usersId":{"type":"integer","format":"int32"}}},"UsersRequestDto":{"type":"object","properties":{"nickname":{"type":"string"},"password":{"type":"string"},"description":{"type":"string"}}},"UsersResponseDto":{"type":"object","properties":{"usersId":{"type":"integer","format":"int32"},"email":{"type":"string"},"nickname":{"type":"string"},"description":{"type":"string"},"isFollowing":{"type":"boolean"},"imageKey":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"roles":{"type":"array","items":{"type":"string"}}}},"PagingResultDtoUsersListResponseDto":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/UsersListResponseDto"}},"totalCount":{"type":"integer","format":"int64"}}},"UsersListResponseDto":{"type":"object","properties":{"usersId":{"type":"integer","format":"int32"},"nickname":{"type":"string"},"imageKey":{"type":"string"}}},"PagingResultDtoPostResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}},"totalCount":{"type":"integer","format":"int64"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"CommentTreeResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"content":{"type":"string"},"likeCount":{"type":"integer","format":"int32"},"childCount":{"type":"integer","format":"int32"},"isDeleted":{"type":"boolean"},"deletedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int32"},"username":{"type":"string"},"parentCommentId":{"type":"integer","format":"int32"},"isLiked":{"type":"boolean"}}},"PagingResultDtoCommentTreeResponseDto":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/CommentTreeResponseDto"}},"totalCount":{"type":"integer","format":"int64"}}},"CsrfToken":{"type":"object","properties":{"headerName":{"type":"string"},"token":{"type":"string"},"parameterName":{"type":"string"}}}}}}